requirements.txt
# Core ML/DL Frameworks
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
tensorflow>=2.15.0
keras>=3.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.1.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# NLP & LLMs
transformers>=4.35.0
tokenizers>=0.15.0
sentence-transformers>=2.2.0
langchain>=0.1.0
llama-index>=0.9.0

# Model Training & Tracking
pytorch-lightning>=2.1.0
tensorboard>=2.15.0
wandb>=0.16.0
mlflow>=2.9.0

# Data Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

# Jupyter & Development
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.1.0
notebook>=7.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.12.0
flake8>=6.1.0
isort>=5.13.0
mypy>=1.7.0
